simpler build

+++ build.py
@@ -1,6 +1,5 @@
 from fastapi import Request
-from main import templates, photos 
-import random 
+from main import index
 import shutil
 from pathlib import Path 
 
@@ -8,13 +7,15 @@ site_path = Path.cwd() / 'site'
 if not site_path.exists():
     site_path.mkdir(parents=True, exist_ok=True)
 
-def index():
-    request = Request
-    oral_histories, photographs, teaching = random.sample(photos, 3)
-    page = templates.TemplateResponse("index.html", {"request":request,"oral_histories": oral_histories, "photographs":photographs, "teaching":teaching})
+def build_index():
+    page = index(Request)
     (site_path / 'index.html').write_bytes(page.body)
 
 if __name__ == '__main__':
     # copy assets to site 
-    shutil.copytree((Path.cwd() / 'assets'), (site_path / 'assets')) 
-    index()
+    if not (site_path / 'assets').exists():
+        shutil.copytree((Path.cwd() / 'assets'), (site_path / 'assets')) 
+    else:
+        shutil.rmtree((site_path / 'assets'))
+        shutil.copytree((Path.cwd() / 'assets'), (site_path / 'assets')) 
+    build_index()
